Error handling & refactoring student serv.

+++ deliverables/yres_scheduler/yres_scheduler_backend/api/controllers/studentController.js
@@ -1,7 +1,8 @@
 const c = require('config');
 const studentService = require('../services/studentService');
-const {STATUS_CODES} = require('../entities/ServiceErrors');
+const {StudentServiceError, STATUS_CODES} = require('../entities/ServiceErrors');
 const logger = require('../../logger');
+
 /**
  * Retrieves all students.
  * @param {Object} req - The request object.
@@ -10,30 +11,19 @@ const logger = require('../../logger');
  */
 async function getAllStudents(req, res) {
     
-    try {
-        const all_students = await studentService.getAllStudents();
-        if (all_students?.result === false) {
+    const all_students = await studentService.getAllStudents();
+
+    res.status(STATUS_CODES.SUCCESS);
+
+    return {
+        result: all_students.map((student) => { 
             return {
-                status: STATUS_CODES.FAILED,
-                error: all_students.error
+                ...student,
+                friend_ids: student.getFriendIds(),
+                enemy_ids: student.getEnemyIds()
             };
-        }
-        return {
-            status: STATUS_CODES.SUCCESS,
-            result: all_students.map((student) => { 
-                return {
-                    ...student,
-                    friend_ids: student.getFriendIds(),
-                    enemy_ids: student.getEnemyIds()
-                };
-            })
-        };
-    } catch (error) {
-        return {
-            status: STATUS_CODES.FAILED,
-            error: error.message
-        };
-    }
+        })
+    };
     
 }
 
@@ -45,26 +35,23 @@ async function getAllStudents(req, res) {
  */
 function getStudentById(req, res) {
 
-    try {
-        const student_id = req.body.student_id;
+    const student_id = req.body.student_id;
 
-        const student = studentService.getStudentById(student_id);
-        if (student?.result === false) {
-            return {
-                status: STATUS_CODES.FAILED,
-                error: student.error
-            };
-        }
-        return {
-            result: student,
-            status: STATUS_CODES.SUCCESS
-        };
-    } catch (error) {
-        return {
-            status: STATUS_CODES.FAILED,
-            error: error.message
-        };
+    // Check paramaters are valid
+    if (!student_id) {
+        throw new StudentServiceError(
+            `Invalid paramaters provided for request`,
+            STATUS_CODES.INVALID
+        );
     }
+
+    const student = studentService.getStudentById(student_id);
+
+    res.status(STATUS_CODES.SUCCESS);
+    
+    return {
+        result: student
+    };
 }
 
 /**
@@ -74,26 +61,22 @@ function getStudentById(req, res) {
  * @returns {Object} - An object containing a student.
  */
 function getStudentByUiId(req, res) {
-    try {
-        const stduent_ui_id = req.body.stduent_ui_id;
+    const student_ui_id = req.params.student_id;
 
-        const student = studentService.getStudentByUiId(stduent_ui_id);
-        if (student?.result === false) {
-            return {
-                status: STATUS_CODES.FAILED,
-                error: student.error
-            };
-        }
-        return {
-            status: STATUS_CODES.SUCCESS,
-            result: student
-        };
-    } catch (error) {
-        return {
-            status: STATUS_CODES.FAILED,
-            error: error.message
-        };
+    if (!student_ui_id) {
+        throw new StudentServiceError(
+            `Invalid paramaters provided for request`,
+            STATUS_CODES.INVALID
+        );
     }
+
+    const student = studentService.getStudentByUiId(student_ui_id);
+
+    res.status(STATUS_CODES.SUCCESS);
+
+    return {
+        result: student
+    };
 }
 
 /**
@@ -106,25 +89,21 @@ async function createStudent(req, res) {
     
     const student = req.body;
 
-    try {
-        const status = await studentService.createStudent(student);
-        
-        if (status?.result === false) {
-            return {
-                status: STATUS_CODES.FAILED,
-                error: status.error
-            };
-        } else {
-            return {
-                status: STATUS_CODES.SUCCESS,
-            };
-        }
-    } catch (error) {
-        return {
-            status: STATUS_CODES.FAILED,
-            error: error.message
-        };
+    if (!student) {
+        throw new StudentServiceError(
+            `Invalid paramaters provided for request`,
+            STATUS_CODES.INVALID
+        );
     }
+
+    const status = await studentService.createStudent(student);
+
+    res.status(STATUS_CODES.SUCCESS);
+    
+    return {
+        status: status
+    };
+
 }
 
 /**
@@ -161,27 +140,22 @@ async function createStudentsFromList(req, res) {
  */
 async function editStudentById(req, res) {
 
-    try {
-        const student = req.body;
+    const student = req.body;
 
-        const status = await studentService.editStudentById(student);
-
-        if (status?.result === false) {
-            return {
-                status: STATUS_CODES.FAILED,
-                error: status.error
-            };
-        } else {
-            return {
-                status: STATUS_CODES.SUCCESS,
-            };
-        }
-    } catch (error) {
-        return {
-            status: STATUS_CODES.FAILED,
-            error: error.message
-        };
+    if (!student) {
+        throw new StudentServiceError(
+            `Invalid paramaters provided for request`,
+            STATUS_CODES.INVALID
+        );
     }
+
+    const status = await studentService.editStudentById(student);
+
+    res.status(STATUS_CODES.SUCCESS);
+
+    return {
+        status: status
+    };
 }
 
 /**
@@ -191,27 +165,22 @@ async function editStudentById(req, res) {
  * @returns {Object} - An object containing a status message.
  */
 async function deleteStudentById(req, res) {
-    try {
-        const student_ui_id = req.body.student_ui_id;
+    const student_ui_id = req.params.student_id;
 
-        const status = await studentService.deleteStudentById(student_ui_id);
-
-        if (status?.result === false) {
-            return {
-                status: STATUS_CODES.FAILED,
-                error: status.error
-            };
-        } else {
-            return {
-                status: STATUS_CODES.SUCCESS,
-            };
-        }
-    } catch (error) {
-        return {
-            status: STATUS_CODES.FAILED,
-            error: error.message
-        };
+    if (!student_ui_id) {
+        throw new StudentServiceError(
+            `Invalid paramaters provided for request`,
+            STATUS_CODES.INVALID
+        );
     }
+
+    const status = await studentService.deleteStudentById(student_ui_id);
+
+    res.status(STATUS_CODES.SUCCESS);
+
+    return {
+        status: status
+    };
 }
 
 module.exports = {
